pyproject.toml
[project]
name = "react-rag-agent"
version = "0.1.0"
description = "A local ReAct RAG agent powered by Ollama, LangGraph, and ChromaDB"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.6",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]

[project.scripts]
react-rag-agent = "react_rag_agent.main:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 100
target-version = "py312"
